moves getChanges to utils w/ tests
